Add new graph structure (#1404)
Authors:
  - https://github.com/Iroy30

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)
  - Brad Rees (https://github.com/BradReesWork)
  - Joseph Nke (https://github.com/jnke2016)
  - Alex Fender (https://github.com/afender)

URL: https://github.com/rapidsai/cugraph/pull/1404
