+++ python/cugraph/components/connectivity_wrapper.pyx
@@ -22,7 +22,7 @@ from cugraph.structure import utils_wrapper
 from cugraph.structure import graph_primtypes_wrapper
 from libc.stdint cimport uintptr_t
 from cugraph.structure.symmetrize import symmetrize
-from cugraph.structure.graph import Graph as type_Graph
+from cugraph.structure.graph_classes import Graph as type_Graph
 import cudf
 import numpy as np
