+++ python/cugraph/link_analysis/pagerank.py
@@ -12,7 +12,7 @@
 # limitations under the License.
 
 from cugraph.link_analysis import pagerank_wrapper
-from cugraph.structure.graph import null_check
+from cugraph.structure.graph_classes import null_check
 import cugraph
