+++ python/cugraph/link_prediction/jaccard.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2021, NVIDIA CORPORATION.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -13,9 +13,8 @@
 
 import pandas as pd
 import cudf
-from cugraph.structure.graph import Graph
+from cugraph.structure.graph_classes import Graph, null_check
 from cugraph.link_prediction import jaccard_wrapper
-from cugraph.structure.graph import null_check
 from cugraph.utilities import check_nx_graph
 from cugraph.utilities import df_edge_score_to_dictionary
