+++ python/cugraph/structure/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2021, NVIDIA CORPORATION.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,7 +11,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from cugraph.structure.graph import Graph, DiGraph, MultiGraph, MultiDiGraph
+from cugraph.structure.graph_classes import (Graph,
+                                             DiGraph,
+                                             MultiGraph,
+                                             MultiDiGraph,
+                                             BiPartiteGraph,
+                                             BiPartiteDiGraph)
+from cugraph.structure.graph_classes import (is_weighted,
+                                             is_directed,
+                                             is_multigraph,
+                                             is_bipartite,
+                                             is_multipartite)
 from cugraph.structure.number_map import NumberMap
 from cugraph.structure.symmetrize import symmetrize, symmetrize_df , symmetrize_ddf
 from cugraph.structure.convert_matrix import (from_edgelist,
