+++ python/cugraph/structure/convert_matrix.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2021, NVIDIA CORPORATION.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -18,7 +18,7 @@
 import cudf
 import dask_cudf
 
-from cugraph.structure.graph import DiGraph, Graph
+from cugraph.structure.graph_classes import DiGraph, Graph
 
 # optional dependencies used for handling different input types
 try:
