+++ python/cugraph/structure/symmetrize.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2021, NVIDIA CORPORATION.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from cugraph.structure import graph as csg
+from cugraph.structure import graph_classes as csg
 import cudf
 import dask_cudf
 
@@ -201,8 +201,12 @@ def symmetrize(source_col, dest_col, value_col=None, multi=False,
         csg.null_check(source_col)
         csg.null_check(dest_col)
     if value_col is not None:
-        weight_name = "value"
-        input_df.insert(len(input_df.columns), "value", value_col)
+        if isinstance(value_col, cudf.Series):
+            weight_name = "value"
+            input_df.insert(len(input_df.columns), "value", value_col)
+        elif isinstance(value_col, cudf.DataFrame):
+            input_df = cudf.concat([input_df, value_col], axis=1)
+
     output_df = None
     if type(source_col) is dask_cudf.Series:
         output_df = symmetrize_ddf(
@@ -211,11 +215,17 @@ def symmetrize(source_col, dest_col, value_col=None, multi=False,
     else:
         output_df = symmetrize_df(input_df, "source", "destination", multi,
                                   symmetrize)
-
     if value_col is not None:
-        return (
-            output_df["source"],
-            output_df["destination"],
-            output_df["value"],
-        )
+        if isinstance(value_col, cudf.Series):
+            return (
+                output_df["source"],
+                output_df["destination"],
+                output_df["value"],
+            )
+        elif isinstance(value_col, cudf.DataFrame):
+            return (
+                output_df["source"],
+                output_df["destination"],
+                output_df[value_col.columns],
+            )
     return output_df["source"], output_df["destination"]
