+++ python/cugraph/traversal/traveling_salesperson.py
@@ -12,7 +12,7 @@
 # limitations under the License.
 
 from cugraph.traversal import traveling_salesperson_wrapper
-from cugraph.structure.graph import null_check
+from cugraph.structure.graph_classes import null_check
 import cudf
